minor cleanups to config state
