+++ .changeset/ten-donkeys-grab.md
@@ -1,0 +1,4 @@
+---
+'@finos/legend-application-studio': patch
+'@finos/legend-query-builder': patch
+---
