+++ packages/legend-application-studio/src/stores/editor-state/project-configuration-editor-state/ProjectConfigurationEditorState.ts
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import type { EditorStore } from '../EditorStore.js';
-import { EditorState } from '../editor-state/EditorState.js';
+import type { EditorStore } from '../../EditorStore.js';
+import { EditorState } from '../EditorState.js';
 import {
   action,
   computed,
@@ -33,25 +33,16 @@ import {
   hashArray,
   ActionState,
 } from '@finos/legend-shared';
-import type { EditorSDLCState } from '../EditorSDLCState.js';
+import type { EditorSDLCState } from '../../EditorSDLCState.js';
 import {
   type ProjectConfiguration,
   ProjectStructureVersion,
   UpdateProjectConfigurationCommand,
   UpdatePlatformConfigurationsCommand,
 } from '@finos/legend-server-sdlc';
-import { LEGEND_STUDIO_APP_EVENT } from '../LegendStudioAppEvent.js';
-import {
-  type ProjectDependencyGraphReport,
-  type ProjectDependencyVersionNode,
-  MASTER_SNAPSHOT_ALIAS,
-  ProjectData,
-  ProjectDependencyCoordinates,
-  RawProjectDependencyReport,
-  buildDependencyReport,
-  buildConflictsPaths,
-} from '@finos/legend-server-depot';
-import type { TreeData, TreeNodeData } from '@finos/legend-art';
+import { LEGEND_STUDIO_APP_EVENT } from '../../LegendStudioAppEvent.js';
+import { MASTER_SNAPSHOT_ALIAS, ProjectData } from '@finos/legend-server-depot';
+import { ProjectDependencyEditorState } from './ProjectDependencyEditorState.js';
 
 export enum CONFIGURATION_EDITOR_TAB {
   PROJECT_STRUCTURE = 'PROJECT_STRUCTURE',
@@ -59,68 +50,9 @@ export enum CONFIGURATION_EDITOR_TAB {
   PLATFORM_CONFIGURATIONS = 'PLATFORM_CONFIGURATIONS',
 }
 
-export class DependencyTreeNodeData implements TreeNodeData {
-  value: ProjectDependencyVersionNode;
-  id: string;
-  label: string;
-  childrenIds?: string[] | undefined;
-
-  constructor(id: string, value: ProjectDependencyVersionNode) {
-    this.id = id;
-    this.value = value;
-    this.label = value.id;
-  }
-  isSelected?: boolean | undefined;
-  isOpen?: boolean | undefined;
-}
-
-export const buildDependencyNodeChildren = (
-  parentNode: DependencyTreeNodeData,
-  treeNodes: Map<string, DependencyTreeNodeData>,
-): void => {
-  if (!parentNode.childrenIds) {
-    const value = parentNode.value;
-    const childrenNodes = value.dependencies.map((projectVersion) => {
-      const childId = `${parentNode.id}.${projectVersion.id}`;
-      const childNode = new DependencyTreeNodeData(childId, projectVersion);
-      treeNodes.set(childId, childNode);
-      return childNode;
-    });
-    parentNode.childrenIds = childrenNodes.map((c) => c.id);
-  }
-};
-
-const buildDependencyTreeData = (
-  report: ProjectDependencyGraphReport,
-): TreeData<DependencyTreeNodeData> => {
-  const nodes = new Map<string, DependencyTreeNodeData>();
-  const rootNodes = report.graph.rootNodes.map((versionNode) => {
-    const node = new DependencyTreeNodeData(versionNode.id, versionNode);
-    nodes.set(node.id, node);
-    buildDependencyNodeChildren(node, nodes);
-    return node;
-  });
-  const rootIds = rootNodes.map((node) => node.id);
-  return { rootIds, nodes };
-};
-
-const buildFlattenDependencyTreeData = (
-  report: ProjectDependencyGraphReport,
-): TreeData<DependencyTreeNodeData> => {
-  const nodes = new Map<string, DependencyTreeNodeData>();
-  const rootIds: string[] = [];
-  Array.from(report.graph.nodes.entries()).forEach(([key, value]) => {
-    const id = value.id;
-    const node = new DependencyTreeNodeData(id, value);
-    nodes.set(id, node);
-    rootIds.push(id);
-    buildDependencyNodeChildren(node, nodes);
-  });
-  return { rootIds, nodes };
-};
-
 export class ProjectConfigurationEditorState extends EditorState {
   sdlcState: EditorSDLCState;
+  projectDependencyEditorState: ProjectDependencyEditorState;
   originalProjectConfiguration?: ProjectConfiguration | undefined; // TODO: we might want to remove this when we do change detection for project configuration
   projectConfiguration?: ProjectConfiguration | undefined;
   selectedTab: CONFIGURATION_EDITOR_TAB;
@@ -128,12 +60,7 @@ export class ProjectConfigurationEditorState extends EditorState {
   projects = new Map<string, ProjectData>();
   queryHistory = new Set<string>();
   latestProjectStructureVersion: ProjectStructureVersion | undefined;
-  dependencyReport: ProjectDependencyGraphReport | undefined;
-  dependencyTreeData: TreeData<DependencyTreeNodeData> | undefined;
-  flattenDependencyTreeData: TreeData<DependencyTreeNodeData> | undefined;
-  dependencyTreeReportModal = false;
-  dependencyConflictModal = false;
-  fetchingDependencyInfoState = ActionState.create();
+
   updatingConfigurationState = ActionState.create();
   fetchingProjectVersionsState = ActionState.create();
   associatedProjectsAndVersionsFetched = false;
@@ -152,28 +79,22 @@ export class ProjectConfigurationEditorState extends EditorState {
       associatedProjectsAndVersionsFetched: observable,
       fetchingProjectVersionsState: observable,
       latestProjectStructureVersion: observable,
-      dependencyReport: observable,
-      dependencyTreeReportModal: observable,
-      fetchingDependencyInfoState: observable,
-      dependencyConflictModal: observable,
-      dependencyTreeData: observable.ref,
-      flattenDependencyTreeData: observable.ref,
+      projectDependencyEditorState: observable,
       originalConfig: computed,
       setOriginalProjectConfiguration: action,
-      setDependencyConflictModal: action,
-      clearTrees: action,
       setProjectConfiguration: action,
-      setDependencyTreeReportModal: action,
       setSelectedTab: action,
-      setDependencyTreeData: action,
       fectchAssociatedProjectsAndVersions: flow,
       updateProjectConfiguration: flow,
       updateToLatestStructure: flow,
       updateConfigs: flow,
       fetchLatestProjectStructureVersion: flow,
-      fetchDependencyInfo: flow,
     });
 
+    this.projectDependencyEditorState = new ProjectDependencyEditorState(
+      this,
+      this.editorStore,
+    );
     this.selectedTab = CONFIGURATION_EDITOR_TAB.PROJECT_STRUCTURE;
     this.isReadOnly = editorStore.isInViewerMode;
     this.sdlcState = sdlcState;
@@ -191,31 +112,6 @@ export class ProjectConfigurationEditorState extends EditorState {
 
   setSelectedTab(tab: CONFIGURATION_EDITOR_TAB): void {
     this.selectedTab = tab;
-  }
-
-  setDependencyConflictModal(showModal: boolean): void {
-    this.dependencyConflictModal = showModal;
-  }
-
-  setDependencyTreeReportModal(showModal: boolean): void {
-    this.dependencyTreeReportModal = showModal;
-  }
-
-  setFlattenDependencyTreeData(
-    tree: TreeData<DependencyTreeNodeData> | undefined,
-  ): void {
-    this.flattenDependencyTreeData = tree;
-  }
-
-  clearTrees(): void {
-    this.flattenDependencyTreeData = undefined;
-    this.dependencyTreeData = undefined;
-  }
-
-  setDependencyTreeData(
-    tree: TreeData<DependencyTreeNodeData> | undefined,
-  ): void {
-    this.dependencyTreeData = tree;
   }
 
   get label(): string {
@@ -286,54 +182,6 @@ export class ProjectConfigurationEditorState extends EditorState {
       );
     } finally {
       this.fetchingProjectVersionsState.complete();
-    }
-  }
-
-  *fetchDependencyInfo(): GeneratorFn<void> {
-    try {
-      this.fetchingDependencyInfoState.inProgress();
-      this.dependencyReport = undefined;
-      this.clearTrees();
-      if (this.projectConfiguration?.projectDependencies) {
-        const dependencyCoordinates = (yield flowResult(
-          this.editorStore.graphState.buildProjectDependencyCoordinates(
-            this.projectConfiguration.projectDependencies,
-          ),
-        )) as ProjectDependencyCoordinates[];
-        const dependencyInfoRaw =
-          (yield this.editorStore.depotServerClient.analyzeDependencyTree(
-            dependencyCoordinates.map((e) =>
-              ProjectDependencyCoordinates.serialization.toJson(e),
-            ),
-          )) as PlainObject<RawProjectDependencyReport>;
-        const rawdependencyReport =
-          RawProjectDependencyReport.serialization.fromJson(dependencyInfoRaw);
-        const report = buildDependencyReport(rawdependencyReport);
-        this.dependencyReport = report;
-        this.processReport(report);
-      }
-      this.fetchingDependencyInfoState.complete();
-    } catch (error) {
-      assertErrorThrown(error);
-      this.fetchingDependencyInfoState.fail();
-      this.dependencyReport = undefined;
-      this.editorStore.applicationStore.log.error(
-        LogEvent.create(LEGEND_STUDIO_APP_EVENT.DEPOT_MANAGER_FAILURE),
-        error,
-      );
-    }
-  }
-
-  processReport(report: ProjectDependencyGraphReport): void {
-    this.setDependencyTreeData(buildDependencyTreeData(report));
-    this.setFlattenDependencyTreeData(buildFlattenDependencyTreeData(report));
-    try {
-      report.conflictPaths = buildConflictsPaths(report);
-    } catch (error) {
-      assertErrorThrown(error);
-      this.editorStore.applicationStore.notifyError(
-        `Unable to build conflict paths ${error.message}`,
-      );
     }
   }
